Implementation of the Fuller et al. angular momentum transport
There are two versions, from Fuller et al. 2019 and Fuller & Lu 2022,
respectively.

In both cases, the functions have been taken from the zenodo repositories
associated to these papers (all linked in the README here) and adapted for
the current MESA version (r24.03.1). Specifically this means:

* removing `_cr` from the math functions (since cr_lib is not part of MESA anymore)
* propagating the star `id` from the TSF* function to diffmag and ultimately `sige2`
* adding a `s%` pointer in `sige2`, intialize it with `id` above
  to obtain `s% kap_handle` to obtain the electron conduction opacity

The rest of the code is left untouched (including their comments) to have
a bug-for-bug reproduction of the original work of Jim Fuller and collaborators.

The example folder just calls the function corresponding to the Fuller & Lu 2022
implementation and is kept very minimal.

This is intended to make it easy to test this angular momentum prescriptions.